requirements.txt
mysql-connector-python
flask==3.0.0
pyecharts==2.0.1